# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_acerowl_lockin_backend_internal_models.User:
    properties:
      createdAt:
        type: string
      currentStreak:
        type: integer
      dailyCommitment:
        type: integer
      deletedAt:
        type: string
      email:
        type: string
      goal:
        type: string
      id:
        type: string
      lastSessionDate:
        type: string
      longestStreak:
        type: integer
      onboardingCompleted:
        type: boolean
    type: object
...
  internal_handlers.AuthTokenData:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
  internal_handlers.AuthTokenResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers.AuthTokenData'
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
      success:
        example: false
        type: boolean
    type: object
  internal_handlers.ForgotPasswordReq:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
...
  internal_handlers.LoginReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: supersecret
        type: string
    type: object
...
  internal_handlers.RefreshReq:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
  internal_handlers.RefreshTokenData:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
  internal_handlers.RefreshTokenResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers.RefreshTokenData'
      success:
        example: true
        type: boolean
    type: object
  internal_handlers.RegisterReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: supersecret
        type: string
    type: object
...
  internal_handlers.SuccessResponse:
    properties:
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.UpdateMeReq:
    properties:
      daily_commitment:
        example: 15
        type: integer
      goal:
        example: Academic Excellence
        type: string
      onboarding_completed:
        example: true
        type: boolean
    type: object
  internal_handlers.UpdateStatusReq:
    properties:
      status:
        type: string
    type: object
...
  internal_handlers.UserResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.User'
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: LockIn Support
    url: https://github.com/acerowl/lockin
  description: Adaptive learning backend — quiz generation, sessions, and progress
    tracking.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: LockIn API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset token to the user's email (simulated)
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ForgotPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Reset request processed
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access + refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens and user profile
          schema:
            $ref: '#/definitions/internal_handlers.AuthTokenResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access + refresh token
        pair
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RefreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: New token pair
          schema:
            $ref: '#/definitions/internal_handlers.RefreshTokenResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns access + refresh tokens
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens and user profile
          schema:
            $ref: '#/definitions/internal_handlers.AuthTokenResponse'
        "400":
          description: Validation or registration error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
  /modules/status/{id}:
    post:
      consumes:
      - application/json
      description: Updates the status of a specific module
      parameters:
      - description: Module UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateStatusReq'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update module status
      tags:
      - modules
...
  /users/me:
    get:
      description: Returns the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/internal_handlers.UserResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Updates onboarding fields (goal, daily_commitment, onboarding_completed)
        for the authenticated user
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateMeReq'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile
          schema:
            $ref: '#/definitions/internal_handlers.UserResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"